semrush_mailer/requirements.txt
selenium>=4.0.0
webdriver-manager>=3.8.0
requests>=2.28.0
python-dotenv>=0.20.0
jinja2>=3.0.0
pillow>=9.0.0 